Modify quesadilla.html capitalization, wording, headings, and spacing
I changed the size of the description heading, and the description
itself as well. To the ingredients list, I capitalized every word
and added space between each entry in the list. Lastly, in the Steps
list I took out the breaks that, instead of allowing each entry
to run on, broke each single entry over multiple lines. I thought
it looked nicer to just allow it to run on instead. I also changed
the "Assembly" header to "Finishing it off....", and I added a link
to the bottom of the page to link back to the index page. I'll probably
fidget more with the wording later.