Modify heading sizes, wording, capitalization, spaces and add link to index
I changed the description again (and will probably change it again
later), and took out all the 'ofs' in the ingredient list.
Changed 'Tablespoon' to Tbsp and 'teaspoon' to tsp.
Changed a few of the Toppings wordings, and changed "The Rest" under
Steps to "Two Hours Later..."I changed some of the wording of the steps,
took out and added spaces, added comments on what to change later,
and added a link at the bottom of the page to the index.html page