Modify spacing, heading size, capitalization, etc in son-in-law-eggs.html
I added and took out spaces, and fixed the capitalization of the steps
in order to match the other recipe pages. I changed Tablespoon to Tbps and
teaspoon to tsp, and changed the wording of some of the notes and
italicized the notes as well. I changed some of the wording of the
ingredients list and steps, and I added a link to the index page
at the bottom of the son-in-law-eggs.html page. Lastly, I added
periods to the ends of many of the 'steps' for consistency.